+++ CHANGELOG.md
@@ -5,11 +5,11 @@ Changes to Calva.
 - [Support for custom project/workflow commands](https://github.com/BetterThanTomorrow/calva/issues/281)
 
 ## Being distilled
-- [Add pretty print mode](https://github.com/BetterThanTomorrow/calva/issues/327)
 
 ## [Unreleased]
 - [Stop writing results from **Evaluate to Comment** to output pane](https://github.com/BetterThanTomorrow/calva/issues/347)
 - [Add command for evaluating top level form as comment](https://github.com/BetterThanTomorrow/calva/issues/349)
+- [Add pretty print mode](https://github.com/BetterThanTomorrow/calva/issues/327)
 
 ## [2.0.40] - 25.09.2019
 - [Add command for connecting to a non-project REPL](https://github.com/BetterThanTomorrow/calva/issues/328)