+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changes to Calva.
 
 ## [Unreleased]
 - [Adding selected calva commands to the editors context menu](https://github.com/BetterThanTomorrow/calva/issues/338)
+
+## [2.0.41] - 28.09.2019
 - [Add pretty print mode](https://github.com/BetterThanTomorrow/calva/issues/327)
 - [Add command for evaluating top level form as comment](https://github.com/BetterThanTomorrow/calva/issues/349)
 - [Stop writing results from **Evaluate to Comment** to output pane](https://github.com/BetterThanTomorrow/calva/issues/347)